// src/app/shared/models/application.model.ts
export interface Application {
    id: string;
    email: string;
    applyfor: string;
    dateapply: Date;
    heardabouts: string;
    firstName: string;
    lastName: string;
    middleName: string;
    number: string;
    street: string;
    city: string;
    county: string;
    state: string;
    zipcode: string;
    phone: string;
    ssn: string;
     q1y: boolean;
     q1n: boolean;
     q2y: boolean;
     q2n: boolean;
     ifyes: Date;
     q3y: boolean;
     q3n: boolean;
     ifemployedwhenandwhere: string;
     q4y: boolean;
     q4n: boolean;
     q5y: boolean;
     q5n: boolean;
     q6y: boolean;
     q6n: boolean;
     availabetowork: Date;
     fulltime: boolean;
     parttime: boolean;
     shiftwork: boolean;
     temporary: boolean;
     q7y: boolean;
     q7n: boolean;
     q8y: boolean;
     q8n: boolean;
     q9y: boolean;
     q9n: boolean;
     ifconvited: string;
     elementaryname: string;
     hightschoolname: string;
     collegename: string;
     professionalname: string;
     yearscompleted1: number;
     yearscompleted2: number;
     yearscompleted3: number;
     yearscompleted4: number;
     diplomadegree1: boolean;
     diplomadegree2: boolean;
     diplomadegree3: boolean;
     diplomadegree4: boolean;
     courses: string;
     trainings: string;
     honors: string;
     moreinfo: string;
     speakfluent: string;
     speakgood: string;
     speakfail: string;
     readfluent: string;
     readgood: string;
     readfail: string;
     writefluent: string;
     writegood: string;
     writefail: string;
     emergencyname: string;
     emergencyaddress: string;
     emergencynumber: string;
     usamilitaryyes: boolean;
     usamilitaryno: boolean;
     usamilataryexplain: string;
     employer1name: string;
     employer2name: string;
     employer3name: string;
     employer4name: string;
     employer5name: string;
     employer1address: string;
     employer2address: string;
     employer3address: string;
     employer4address: string;
     employer5address: string;
     from1: string;
     from2: string;
     from3: string;
     from4: string;
     from5: string;
     to1: string;
     to2: string;
     to3: string;
     to4: string;
     to5: string;
     wperformed1: string;
     wperformed2: string;
     wperformed3: string;
     wperformed4: string;
     wperformed5: string;
     employer1number: string;
     employer2number: string;
     employer3number: string;
     employer4number: string;
     employer5number: string;
     start1: string;
     start2: string;
     start3: string;
     start4: string;
     start5: string;
     final1: string;
     final2: string;
     final3: string;
     final4: string;
     final5: string;
     jobtitle1: string;
     jobtitle2: string;
     jobtitle3: string;
     jobtitle4: string;
     jobtitle5: string;
     supervisor1: string;
     supervisor2: string;
     supervisor3: string;
     supervisor4: string;
     supervisor5: string;
     reasonforleaving1: string;
     reasonforleaving2: string;
     reasonforleaving3: string;
     reasonforleaving4: string;
     reasonforleaving5: string;
     skillandqualifications: string;
     signatureofapplicant: string;
     datesignature: string;

}